Irix6.4 (with 7.1 compilers)
Subject: Irix6.4 (with 7.1 compilers) passes 100% with patch

>>>>> "John" == John Stoffel <jfs@fluent.com> writes:

John> I had to apply the following patch to the Irix_6_4.sh hints file to
John> get it to compile under Irix 6.4 with the version 7.1 compilers
John> installed.  Thanks to Hallvard Furuseth for pointing out that the
John> pp_sys_cflags line _has_ to be in the left most column.  This leads me
John> to believe that the following hints files will have problems:

Ooops!  I messed up on the patch for Irix_6_2.sh, so here's the
entire patch again to fix the problem as reported under Irix 6.[24].
I don't have a 6.3 machine to test with right now, since I have to
install the compilers first.  More when I get a chance.

p5p-msgid: <199702130238.VAA24468@jfs.Fluent.COM>