Pod::Text fixes
Here are some bug fixes for Pod::Text.

Is it a bug that text given to =for and =begin doesn't get indented?  It
seems like a bug to me, but I imagine one could argue the other way so I
wanted to test opinion before fixing it.

p5p-msgid: <350.856634588@eeyore.ibcinc.com>