Clean up and document API for hashes
On Tue, 25 Feb 1997 09:01:57 EST, Chip Salzenberg wrote:
>I've had 99.44% good reports about _90.  I'd love to go public
>immediately, but Sarathy's perl_call fix is just too valuable to pass
>up, and there's a taint test in the works.  So please don't tell
>anyone that _90 is the first public beta; that will likely be _91.

FWIW, the Beta Zvoid is alive and kicking on Linux 2.0.28.

You might want to put this in too.  (The patches to hv.h are mostly
to an unused macro that will be useful for extension writers).

p5p-msgid: <199702251824.NAA14859@aatma.engin.umich.edu>