Don't count on 'trap 0' inside () in shell script
When building 5.003_91, I noticed that Configure leaves a load of files lying 
around in /tmp, for example exp21743c, etc.  These appear to contain lists of 
symbols.  A bit of digging reveals that the offender is the perl_exp.SH 
script.  This contains a bit of code of the form

        (
        trap "rm -f ...." 0
        ...
        )

I guess that when the sub-shell invoked by the ( ... ) construct exits, the 
trap 0 should be taken, and the files removed.  Unfortunately, I have noticed 
that on at least 2 systems (UnixWare & Unisys SVR4) this doesn't happen.  The 
following patch fixes it.

p5p-msgid: memo.147326@cix.compulink.co.uk