Exporting UNIVERSAL::can
UNIVERSAL allows the export of 'isa' but not 'can'.    Since it
is useful to be able to apply 'can' to non-references (to avoid having
to do redundant tests), I offer the following tiny patch.

This also updates the pod to make it slightly clearer that both
functions can be applied to arbitrary values, not just references.

p5p-msgid: E0w9DwX-0000Zr-00@taurus.cus.cam.ac.uk