C<new SelectSaver $fh> doesn't always restore
This is a bug report for perl from spider@Orb.Nashua.NH.US,


If you're lucky enough to get a signal (such as ALRM) which is handled
with die() while in a SelectSaver->new($fh) call, your previous output
filehandle is not restored.  It violates the basic rule of 'save and
restore' handling, in that it modifies the state it wants to restore
I<before> it has ensured that the prior state will be restored.  Patch:

p5p-msgid: 199703291906.OAA07232@Orb.Nashua.NH.US