Update File::Path
>>>>> Chip Salzenberg writes:

 > According to Andreas Koenig:
>> I'd prefer to have consistent semantics on all platforms. My patch
>> treats the third parameter ($safe) in a way that matches the current
>> description in the pods better. This means, on all systems a chmod +rw
>> is tried before removing a file or directory unless the $safe
>> parameter is specified.

 > No, that's not useful.  UNIX systems pay *no* attention to the
 > permissions of "x" when unlinking "/y/x"; all that matter are the
 > permissions of "/y".

Ouch. Too-Quick-oh. <Blush>

Sure thing is, I want to chmod 777 always for _directories_ before I
remove them, because they may contain subdirectories that need to be
removed recursively. So I want them both read- and writeable. Stupid
thing was that I applied the same idea to files.

>> In addition File::Path becomes strict clean.

 > Well, that would be useful.  Could you separate that part of the patch
 > from the always-chmod part?

I redid it. This time with a test that fails with current perl and
works with my patch in place.

p5p-msgid: 199704151401.QAA02556@anna.in-berlin.de