Support shared libperl on AIX
The current distribution will not build a shared libperl on AIX. 

Currently the Makefile generated uses the same ld options for building
a shared libperl and dynamicly loaded extentions.  AIX's compiler
requires that you tell it explicitly what symbols to export/import. So,
it needs a separate set of LD options for each.

On top of all this, AIX doesn't seem to have an equivilent to
LD_LIBRARY_PATH. So, the directories to search during runtime must be
explicitly compiled in. In the following patch I compiled in both the
compilation directory (for miniperl and testing) and the installation
directory. I wish someone could suggest something cleaner, but it works.

p5p-msgid: 199704270131.UAA51426@icd.cc.purdue.edu