Building Perl on AIX 4+ with shared libraries and dynamic loading
To whom it may concern,

I'm in the process of compiling/installing Perl 5.004_01 (yes, I know, I
should have done it ages ago :) for AIX 4.1.x and I found that the
problems I reported in November 96 (search Dejanews for article
<567m0g$g04@sifon.cc.mcgill.ca> in comp.lang.perl.misc) are still present,
but not the solutions. In essence, the problem results from a change in
the default alignment of the text, data and loader sections in the shared
libraries and executables. The only modification to the source are in
hints/aix.sh. I include the unified diff below. 

Hoping you find this babble of some use,

p5p-msgid: Pine.SGI.3.91.971022084517.17052F-100000@nazgul.physics.mcgill.ca