Irix hint update
Hopefully, this is the final set of patches for the IRIX hints file
hints/irix_6.sh.  This is rather a multi-person effort, Jarkko and
John Stoffel wrote the major changes (and changed the default to
-n32), I made the last tweaks so that -o32 still works without
errors on the various compiler releases.  I've tested on IRIX-6.2,
IRIX64-6.2, IRIX64-6.4 and IRIX-6.5alpha.

p5p-msgid: yd83es0fp57.fsf@hoshi.engr.sgi.com