Tweaks for perldelta
Here are a few more doc fixes.

The last one merely reformats a paragraph, so that a C<...> construct
is no longer spread across a line boudary.  The perlpod page suggests
this is allowed, but does not actually say so.  However, pod2html can't
handle it and all other instaces have been fixed recently (which would
suggest that it' not allowed).

p5p-msgid: 199705102346.BAA17300@mail.euronet.nl