Binary installers for Perl modules
The following patch makes it possible to trivially write a binary
installer for a module on a given platform.

This still leaves a question on binary uninstaller (on platforms where
there is a standard database of installed software, so it may be easy
to hook into it).  Why is the uninstall target of Makefiles disabled?

Enjoy,

p5p-msgid: 199707210006.UAA06165@monk.mps.ohio-state.edu