op/groups test fails on Linux (groups in /bin)
The test op/groups checks for groups in /usr/bin and /usr/ucb.
This fails on some Linux distributions that has groups in /bin.

A better test would be to just run groups (take whatever is in the
path) and see if it outputs a list that includes the current
group.  Possibly you could prepend /bin:/usr/bin:/usr/ucb to the
current path.  Something like this.

p5p-msgid: 199705191120.NAA08130@ostrich.gaia.swipnet.se