[BUG:PATCH] Missing semicolon message wrong in perldiag
This error message is different between the source and perldiag.pod, so
that "use diagnostics;" doesn't pick it up.

I have moved the message to be in the correct collating order,
_ignoring_ the initial TAB.    Is this the right thing to do?
Is the ordering merely for the convenience of humans, or are there
programs which exploit it?    I note that "use diagnostics;" doesn't.

Warning:  This patch inserts a TAB into perldiag  -  make sure it stays
          as a TAB.

p5p-msgid: E0welEn-0002vT-00@taurus.cus.cam.ac.uk