Win32 sitelib intuition from DLL location
There is code for win32 that determines archlib from the location
of PERL.DLL.  This is only 50% effective, as any sitelib at
the same location is not found.

This simple patch adds that feature.  It is needed for the win32
binary distribution to work in a relocatable fashion.

[editor's note: perhaps only partially applied]

p5p-msgid: 199706231647.MAA23260@aatma.engin.umich.edu