win32 docs and runperl.bat
This patch fills in some gaps in the docs, and adds
runperl.bat.  The fix to pl2bat makes it so that
a #!perl line is always available, so that "perl -x"
always works on a batific file.

This goes over my previous win32 patches (esp. the exec
patch and the pl2bat patch).  I'm going to post a
consolidated win32 patch here soon, never fear.

p5p-msgid: 199707070446.AAA29560@aatma.engin.umich.edu