Easier TCP stack selection for VMS
Here's a patch for [.vms]descrip.mms to make it a touch easier to use UCX
sockets instead of socketshr ones. (Use /Macro= DECC_SOCKETS=1 or
SOCKETSHR_SOCKETS=1 to select the stack)

It also skips the ?2P programs when compiling with __DEBUG__ support, as MMK
gets real unhappy up if you try. This is an evil hack, though, as the real
answer is to fix things so they work either way.

Hopefully Eudora hasn't word wrapped this...

p5p-msgid: 3.0.1.32.19970624151939.00994490@stargate.lbcc.cc.or.us